+++ CHANGELOG.md
@@ -13,7 +13,7 @@ since 0.4.1, and this project adheres to [Semantic Versioning](https://semver.or
 
 ### Development changes
 
-- Introduce a tool for vendoring GitHub assets ([#1146](https://github.com/freedomofpress/dangerzone/issues/1146))
+- Vendor GitHub assets using the `assets` tool ([#1146](https://github.com/freedomofpress/dangerzone/issues/1146))
 
 ## [0.9.0](https://github.com/freedomofpress/dangerzone/compare/v0.9.0...0.8.1)